requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
psutil==5.9.8
requests==2.32.4
